requirements.txt
pytest
boto3
psycopg2-binary